# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/daily-selection": {
      "get": {
        "operationId": "DailySelectionController_findToday",
        "parameters": [
          {
            "name": "modeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DailySelection"
        ]
      }
    },
...
    "/daily-selection/latest": {
      "get": {
        "operationId": "DailySelectionController_getLatest",
        "parameters": [
          {
            "name": "modeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DailySelection"
        ]
      }
    },
...
    "/daily-selection/{modeId}": {
      "get": {
        "operationId": "DailySelectionController_getSelectionByMode",
        "parameters": [
          {
            "name": "modeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DailySelection"
        ]
      }
    },
    "/daily-selection/calendar": {
      "get": {
        "operationId": "DailySelectionController_calendar",
        "parameters": [
          {
            "name": "modeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DailySelection"
        ]
      }
    },
    "/daily-selection/by-date": {
      "get": {
        "operationId": "DailySelectionController_byDate",
        "parameters": [
          {
            "name": "modeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DailySelection"
        ]
      }
    },
...
    "/plays/start": {
      "post": {
        "operationId": "PlaysController_start",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartPlayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Plays"
        ]
      }
    },
    "/plays/{playId}/guess": {
      "post": {
        "operationId": "PlaysController_guess",
        "parameters": [
          {
            "name": "playId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Plays"
        ]
      }
    },
    "/plays/{playId}/attempts": {
      "get": {
        "operationId": "PlaysController_listAttempts",
        "parameters": [
          {
            "name": "playId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Plays"
        ]
      }
    },
    "/plays/progress/{modeConfigId}": {
      "get": {
        "operationId": "PlaysController_getProgress",
        "parameters": [
          {
            "name": "modeConfigId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Plays"
        ]
      }
    },
    "/plays/{playId}/progress": {
      "get": {
        "operationId": "PlaysController_getPlayProgress",
        "parameters": [
          {
            "name": "playId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Plays"
        ]
      }
    },
...
  },
  "info": {
    "title": "API of Dreams",
    "description": "Documentação da API of Dreams",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "StartPlayDto": {
        "type": "object",
        "properties": {}
      },
      "GuessDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}